move relay files around
